clean lại code, thêm check alert xác nhận khi thao tác

+++ app/src/main/java/com/example/quanlydatlich/ui/ProfileActivity.java
@@ -67,17 +67,18 @@ public class ProfileActivity extends AppCompatActivity {
         });
 
         btnLogout.setOnClickListener(v -> {
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.clear();
-            editor.apply();
+            new androidx.appcompat.app.AlertDialog.Builder(this).setTitle("Đăng xuất").setMessage("Bro muốn thoát tài khoản hiện tại à?").setPositiveButton("Đăng xuất", (dialog, which) -> {
+                SharedPreferences.Editor editor = sharedPreferences.edit();
+                editor.clear();
+                editor.apply();
 
-            Toast.makeText(this, "Đã đăng xuất thành công!", Toast.LENGTH_SHORT).show();
+                Toast.makeText(this, "Đã đăng xuất thành công!", Toast.LENGTH_SHORT).show();
 
-            Intent intent = new Intent(ProfileActivity.this, HomeActivity.class);
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-            startActivity(intent);
-            finish();
+                Intent intent = new Intent(ProfileActivity.this, HomeActivity.class);
+                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                startActivity(intent);
+                finish();
+            }).setNegativeButton("Ở lại", null).show();
         });
     }
-
 }
